Add files via upload
This is a reworked version of the perspective_distortion_calc. The originial was the plot and test verseion, were a view more angles were calculated for test purposes. 
I removed these angles as well as some old unused variables and added more comments.  

The namining of vairables and the script in general is very misleading. The word "distortion" isnt actually what is happening it would be closer to "parralax" or something like that. Inside the code the naming could be redone as well.