Added Image Velocity Bisect
I now added the "Velocity Bisect" Function as it is build into the "Satellite_Detection_Loop". This is just a side by side  comparison of the two algorihims, they are similar but the second one uses the corrcted angle and the altitude of the foot point, wich is how it would be done with a real image.