More Comments and Renamed some Variables
Added some Comments and in the Perspective correction part there were some variables names ..._declination. Theclination is the wrong word.
I needed to say elevation, because I am talking about horizontal coordinates. The Variable tilted_elevation or previous (tilted_declination) is basically just the angle between the two lines going from the center of the satellite to the zenith and the line going from the foot point to the zenith. 
The reason it is called elevation is because theoretically the "perspective_correction" function is only able to reconstruct the apperance of satellites flying right above the observer, were the passed angle wouldnt be the "tilted elevation" it would be the normal elevation of the satellite in the sky. In other words it is the case were the foot point (were the satellite is the coseset) is above the observer.  In reality basically no satellite passes exactly overhead, so we have to reconstruct that angle, wich is the angle between the foot point and the satellite, only that in this case the footpoint isnt overhead.